[SGBT] option to disable stochastic gradient
The stochastic gradient calculation is based on a random
subsampling of the traning fold. The stochastic gradient
algorithm can be beneficial relative to the calculation
of the loss function's gradient over the entire training
fold because it can, in principle, be faster, and because
it provides a crude form of regularization. However, it
introduces noise which can cause the model to underfit.
Hence, it is important to be able to turn it off.