[logistic] bugfix for y_array_of_feature
Previously all y values in the context of logistic regression where +1.0
or -1.0; hence, classify_float would return FP_normal, and any other
fpclass value would be invalid. Now we allow any legitimate probability
value, so FP_zero and FP_subnormal in particular are acceptable. We
still need to fail when we find y values that are not valid
probabilities: nans, infinities, and values outside the [0.0, 1.0]
range.