[learn] allow max_trees = 0
The intention is to output the mean model only. However, logic in Sgbt
needs to change to check the exit condition from the learning loop at
the entrance of the loop rather than the exit for this to take effect.
Currently, the algorithm will learn the first tree notwithstanding a
setting of max_trees = 0.