[Robust] a (more) robust estimator of the conditional mean
WIP: Currently, only the mean model is changed relative to Square. The
initial gamma value is defines as the median of the distribution rather
than the sample mean to to make the starting point robust.

While eventually, this might evolve into an implementation of
Friedman's M-TreeBost, the intention for now is stick with the L2
penalty, but pick the split point that maximizes the t-stat of
the improvement in loss, thereby favoring directions of loss
reduction that generalize better, without actually changing
the loss function per se.